added JWT to Testimonies

+++ server/controllers/testimony.controller.js
@@ -5,11 +5,18 @@ const {
     updateTestimonyService,
     deleteTestimonyService,
 } = require('../services/testimony.services');
+const { User } = require('../db');
+
+const verifyJWT = require('../middlewares/verifyJwt');
 
 const createTestimony = async (req, res) => {
     const { name, comment } = req.body;
-
+    const { uid } = req.user;
+    const user = await User.findOne({ where: { id: uid } });
     try {
+        if (user.role !== 'admin') {
+            return res.status(401).json({ error: 'Acceso no autorizado' });
+        }
         const testimony = await createTestimonyService(name, comment);
         res.json(testimony);
     } catch (error) {
@@ -43,6 +50,7 @@ const getTestimonyById = async (req, res) => {
 const updateTestimony = async (req, res) => {
     const testimonyId = req.params.id;
     const { name, comment } = req.body;
+    const { role } = req.user;
 
     try {
         const testimony = await getTestimonyByIdService(testimonyId);
@@ -50,7 +58,13 @@ const updateTestimony = async (req, res) => {
             return res.status(404).json({ error: 'Testimonio no encontrado' });
         }
 
-        // Actualizar el usuario
+        console.log('role', role);
+        // Verificar si el usuario tiene el rol de administrador
+        if (role !== 'admin') {
+            return res.status(401).json({ error: 'Acceso no autorizado GIL' });
+        }
+
+        // Actualizar el testimonio
         const updatedTestimony = await updateTestimonyService(
             testimonyId,
             name,
@@ -64,6 +78,7 @@ const updateTestimony = async (req, res) => {
 
 const deleteTestimony = async (req, res) => {
     const testimonyId = req.params.id;
+    const { role } = req.user;
 
     try {
         const testimony = await getTestimonyByIdService(testimonyId);
@@ -71,13 +86,18 @@ const deleteTestimony = async (req, res) => {
             return res.status(404).json({ error: 'Testimonio no encontrado' });
         }
 
+        // Verificar si el usuario tiene el rol de administrador
+        if (role !== 'admin') {
+            return res.status(403).json({ error: 'Acceso no autorizado' });
+        }
+
+        // Eliminar el testimonio
         await deleteTestimonyService(testimonyId);
         res.sendStatus(204);
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
 };
-
 module.exports = {
     createTestimony,
     getAllTestimonies,
